requirements.txt
streamlit
streamlit-geolocation
opencv-python-headless
ultralytics
numpy
SpeechRecognition
PyAudio
Pillow
